requirements.txt
tabula-py==2.3.2
pandas==1.3.5
PyPDF2==1.26.0
tkdnd==0.2.1
